requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
autopep8==1.5.7
certifi==2025.4.26
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.2
constantly==23.10.4
crispy-bootstrap5==2025.4
cryptography==44.0.2
daphne==3.0.2
dj-database-url==0.5.0
Django==5.1.6
django-axes==7.1.0
django-crispy-forms==2.4
django-tenants==3.7.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.7.0
pycparser==2.22
pyOpenSSL==25.0.0
python-decouple==3.4
pytz==2021.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.8.2
sqlparse==0.4.2
toml==0.10.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
Unipath==1.1
urllib3==2.4.0
zope.interface==7.2
gunicorn
psycopg2-binary
whitenoise